requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
google-api-core==2.17.0
google-api-python-client==2.117.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
multidict==6.0.5
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-decouple==3.8
requests==2.31.0
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.0
yarl==1.9.4
Flask==3.0.2
